08/06

+++ 08_square.py
@@ -1,5 +1,5 @@
 #Write a program to print square of given number.
-num = int(input("enter a num = " ))  # 6 
+num = float(input("enter a num = " ))  # 6 
 square = num * num  
 print("square of" , num , "=" , square)
